+++ GFX4dGrid.h
@@ -35,7 +35,7 @@ class GFX4dGrid {
     void addSlider(uint16_t colorb, uint16_t colorbp, uint16_t tcolor, uint16_t x, uint16_t y, uint16_t w, uint16_t h, void (*callback)(int,int), uint8_t initial_pos, int id);
     void addButton(uint16_t colorb, uint16_t colorbp, uint16_t tcolor, uint16_t x, uint16_t y, uint16_t w, uint16_t h, String text, void (*callback)(int,int), int textsize, int id);
     int AddLedGroup(uint16_t x, uint16_t y, uint16_t colorb, uint16_t tcolor, int count);
-    void updateOutput(uint16_t val, int nOutput);
+    void updateOutput(int32_t val, int nOutput);
     void loop();
 };
